requirements.txt
# AI Dubbing Pipeline Dependencies

# Temporal Workflow Engine
temporalio==1.5.1

# Google Cloud Storage
google-cloud-storage==2.10.0

# Google Translate API
google-cloud-translate==3.12.1

# Database (Async PostgreSQL)
asyncpg==0.29.0
psycopg2-binary==2.9.7

# Audio Processing
ffmpeg-python==0.2.0

# Speech Processing / VAD (Voice Activity Detection)
torch==2.1.1
torchaudio==2.1.1
silero-vad==5.1.2
librosa==0.10.1
soundfile==0.12.1

# OpenAI API for Whisper Transcription
openai>=1.12.0

# Environment Management
python-dotenv==1.0.0

# Async File Operations
aiofiles==23.2.1

# Data Validation
pydantic==2.5.0

# Logging and Utils
structlog==23.1.0
tenacity==8.2.3

# HTTP Client (for future API integrations)
httpx==0.25.2

# Type Hints (Updated for OpenAI compatibility)
typing-extensions>=4.9.0

# Development Dependencies
pytest==7.4.0
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.0

# Monitoring (optional for production)
prometheus-client==0.19.0
replicate